Consistent usage of file separator.
File.separator had a utility constant defined in multiple modules and
usage was inconsistent.

I have split it into the two usage types - the ones which are dealing
with file paths and the ones which are dealing with regex patterns.
Usages have been made consistent with the type of usage, but this
requires a good review of the code to make sure my interpretation of
each usage is correct. Some of the usages are ambiguous, which I have
left comments on.