Don't walk the directory tree to build the ignored file list.
Walking the directory tree to build the ignored file list is a bad idea
because it ends up performing a huge amount of I/O for files which
aren't being accessed.

Doing just this brings the per-file time for `upload translations`
down to about 10 seconds per file, instead of 20-30 minutes.